mission_control_mcp/requirements.txt
# MissionControlMCP Requirements
# Python 3.11+ required

# MCP SDK
mcp>=1.0.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.0.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Vector Store & Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# NLP & Text Processing
nltk>=3.8.0
scikit-learn>=1.3.0

# Utilities
python-dateutil>=2.8.0
pydantic>=2.0.0
